remove proc_macro_hack, restructure to single crate
