+++ Cargo.toml
@@ -1,16 +1,18 @@
 [package]
 name = "ifmt"
-version = "0.3.3"
+version = "0.3.4"
 authors = ["ct-austin <c@ctaustin.me>"]
 edition = "2018"
-readme = "README.md"
 repository = "https://github.com/ct-austin/ifmt"
 description = "Inline expression interpolation for Rust."
 license = "MIT/Apache-2.0"
 
-[dependencies]
-ifmt-impl = { version = "0.3.3", path = "impl" }
-proc-macro-hack = "0.5"
+[lib]
+proc-macro = true
 
-[workspace]
-members = ["demo", "impl"]
+[dependencies]
+syn = { version = "1.0", features = ["full", "extra-traits"]}
+regex = "1.1"
+lazy_static = "1.4"
+quote = "1.0"
+proc-macro2 = "1.0"
