--- demo/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "demo"
-version = "0.0.0"
-authors = ["ct-austin <c@ctaustin.me>"]
-edition = "2018"
-publish = false
-
-[dependencies]
-ifmt = { version = "0.3.3", path = "../" }
