+++ examples/demo.rs
@@ -1,5 +1,6 @@
+use ifmt::iprintln;
+
 fn main() {
-    use ifmt::iprintln;
     let four = 4;
     iprintln!("four plus four is: " four + 4);
     // four plus four is: 8
@@ -7,4 +8,6 @@ fn main() {
     // here's a hex number: 0xdeadbeef
     iprintln!("here's a debugging value: " Some(four);?);
     // here's a debugging value: Some(4)
+    iprintln!("heres a four: {four}")
+    // heres a four: 4
 }
