--- impl/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "ifmt-impl"
-version = "0.3.3"
-authors = ["ct-austin <c@ctaustin.me>"]
-edition = "2018"
-repository = "https://github.com/ct-austin/ifmt"
-description = "Implementation detail of the ifmt crate."
-license = "MIT/Apache-2.0"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = { version = "1.0", features = ["full", "extra-traits"]}
-regex = "1.1"
-lazy_static = "1.4"
-quote = "1.0"
-proc-macro-hack = "0.5"
-proc-macro2 = "1.0"
